% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coups.R
\name{add_coups}
\alias{add_coups}
\title{Identifica golpes e autogolpes}
\usage{
add_coups(.data, ...)
}
\arguments{
\item{.data}{dados do projeto \emph{"Varieties of Democracy"}.}

\item{...}{Variáveis que devem ser especificadas
\itemize{
\item \code{.year} = série histórica
\item \code{.countries} = países
\item \code{.typology_d} = classificação dicotômica
\item \code{.typology_p} = classificação policotômica
\item \code{.reg_info} = nome do regime, quando ele começa e quando ele termina
\item \code{.reg_end_type} = nome do processo mais importante que resultou no
colapso do regime
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Esta função identifica golpes e autogolpes
}
