% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{carframe}
\alias{carframe}
\title{A DSM containing component dependencies of the car frame as reported by \insertCite{Pandremenos.2011;textual}{EAD} in Figure 6.}
\format{
A DSM with 38 by 38 elements.
}
\usage{
data(carframe)
}
\description{
A DSM containing component dependencies of the car frame as reported by \insertCite{Pandremenos.2011;textual}{EAD} in Figure 6.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
