% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costingSystem.R
\name{costingSystem_ABC}
\alias{costingSystem_ABC}
\title{Calculate the Reported Product Costs (PC_H)}
\usage{
costingSystem_ABC(
  RES_CONS_PAT,
  RC_var_i,
  RC_fix_i = NULL,
  DMD,
  RD = c("random", "size-random", "size-misc", "correl-random", "correl-size"),
  AD = c("big-pool", "indexed-2", "indexed-3"),
  ACP = 10
)
}
\arguments{
\item{RES_CONS_PAT}{The resource consumption pattern matrix.}

\item{DMD}{A demand vector. If no demand vector is supplied, each product is assumed to have a demand of one.}

\item{RD}{The resource driver method to allocate resource costs to activity cost pools. The following methods are available:
\describe{
 \item{random}{Resources are randomly assigned to activity cost pools.}
 \item{size-random}{Focus on the resources with the largest monetary value. First, the ACP largest resources are assigned to individual cost pools. The remaining resources are random-ly distributed across ACP.}
 \item{size-misc}{Focus on the resources with the largest monetary value. First, the ACP-1 largest resources are assigned to individual cost pools. The remaining resources are assigned to a miscellaneous cost pool.}
 \item{correl-random}{First ACP resources are randomly assigned to cost pools. The remaining resources are assigned to the ACP with the highest corre-lation.}
 \item{correl-size}{Such as heuristic four, except the initial assignment is done by resource’s size rather than random choice.}
}}

\item{AD}{The activity driver used in the second stage allocation.
\describe{
 \item{big-pool}{The big-pool heuristic uses the largest resource as a driver for each activity cost pool}
 \item{indexed-N}{More information demanding is the indexed method which builds a composite driver for the N-largest resources. See examples how to use this allocation method.}
}}

\item{ACP}{The number of activity cost pools.}

\item{RC_direct}{The direct resource cost vector.}
}
\value{
Returns the reported product costs vector (PC_H)
}
\description{
This function calculates the reported product costs using ABC product costing systems.
Methods which are used in \insertCite{BLH2011;textual}{EAD} are implimented.
}
\examples{

### Calculate the reported product costs ###
data("exampleEAD")
  EAD <- exampleEAD
  RES_CONS_PAT <- EAD[[4]][[1]]$P$RD
  RC_indirect <- EAD[[4]][[1]]$RC$var + EAD[[4]][[1]]$RC$fix
  DMD <-  EAD[[4]][[1]]$DEMAND
  ACP<-1
  RD<-"random"
  AD<-"indexed-2"

  costingSystem_ABC(RES_CONS_PAT,
                   RC_indirect,
                   DMD,
                   RD=RD,
                   AD=AD,
                   ACP=10)
}
\references{
\insertAllCited{}
}
