% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costingSystem.R
\name{costingSystem_VD}
\alias{costingSystem_VD}
\title{Calculate the Reported Product Costs (PC_H)}
\usage{
costingSystem_VD(
  RES_CONS_PAT,
  RC_direct,
  RC_indirect,
  DMD,
  method = c("PU", "DC")
)
}
\arguments{
\item{RES_CONS_PAT}{The resource consumption pattern matrix.}

\item{RC_direct}{The direct resource cost vector.}

\item{RC_indirect}{The indirect resource cost vector.}

\item{DMD}{A demand vector. If no demand vector is supplied, each product is assumed to have a demand of one.}

\item{method}{The allocation method. The following methods are available:
\describe{
 \item{PU}{Indirect costs are allocated equally across products.}
 \item{DC}{Indirect costs are allocated based on the direct costs.}
}}
}
\value{
Returns the reported product costs vector (PC_H)
}
\description{
This function calculates the reported product costs using a simple volume-based product costing system.
}
\examples{

### Calculate the reported product costs ###
data("exampleEAD")
  EAD <- exampleEAD
  RES_CONS_PAT <- EAD[[4]][[1]]$P$RD
  RC_indirect <- EAD[[4]][[1]]$RC$var + EAD[[4]][[1]]$RC$fix
  RC_direct <- EAD[[4]][[1]]$RC$direct
  DMD <-  EAD[[4]][[1]]$DEMAND

  costingSystem_VD(RES_CONS_PAT,
                   RC_direct,
                   RC_indirect,
                   DMD,
                   method="PU")
}
