% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ProductMix.R
\name{create_ProductMix}
\alias{create_ProductMix}
\title{Create product mix (P_FD)}
\usage{
create_ProductMix(
  N_FR = 7,
  PARAM = 0.05,
  method = c("combination", "random", "PCI", "DNS")
)
}
\arguments{
\item{N_FR}{Number of functional requirements. It is recommend to keep the value between 3-15.
Product mixes with much more than 15 functional requirements require much RAM as all combinations are generated. Therefore, the function is stopped if \code{N_FR>15}}

\item{PARAM}{A numeric value between \code{0<PARAM<=1} where its meaning depend on the selected \code{'method'}.}

\item{method}{Character which specifies the method, used to generate the product mix. There are two options:
\describe{
  \item{combination}{The product mix is created by using the DSM_FD matrix which contains the configuration constraints in conjuntive normal form.}
  \item{random}{Randomly choose \code{PARAM} percentage of products from the product mix. Is is ensured, that each functional requirement is used at least ones.}
  \item{PCI}{Creates a product mix based on a pre defined product line commonality index (PCI). The density must be within the bounds of \code{0<PARAM<=0.5}.}
  \item{DNS}{Creates a product matrix with a given density.}
}}
}
\value{
A list object containing the product mix. This list consists of:
\describe{
  \item{P_FD}{The free product mix matrix}
  \item{P_FD_const}{The constraint product mix matrix}
  \item{DSM_FD}{The DSM_FD matrix containing the entries to constraint the variety.}
  \item{measures}{A named list containing the following measures: number of product variants (N_P), uniform diversification index (D_u),
    local outlier factor (LOF) and Product Variety (PV) }
}
}
\description{
Creates a product mix based on a given number of functional requirements \code{N_FR} and a density \code{PARAM}.
}
\examples{
set.seed(1234)
prodMIX<-create_ProductMix(N_FR=7,PARAM=0.2)
prodMIX$P_FD_const
}
