% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_DMM.R
\name{crt_DMM}
\alias{crt_DMM}
\title{Create a domain mapping matrix (DMM)}
\usage{
crt_DMM(N_src, N_tgt, DMM_PAR, binary = F, upper_Bound = 20, allowZero = T)
}
\arguments{
\item{N_src}{Number of rows for the DMM.}

\item{N_tgt}{Number of columns for the DMM.}

\item{DMM_PAR}{Double within \code{0<=DMM_PAR<=1}. Specifies the desired system design complexity which should be generated.}

\item{binary}{Boolean, default=F. If the output DMM should be an binary or non binary matrix. If \code{binary=TRUE} then the input \code{'upper_Bound'} will be ignored.}

\item{upper_Bound}{Integer, default \code{upper_Bound=20}, specifies the maximum number in the DMM if \code{binary=FALSE}}

\item{allowZero}{Boolean, default \code{allowZero=T}, specifies whether zero columns are allowed in the DMM matrices.}
}
\value{
A list object containing the product mix. This list consists of:
\describe{
  \item{DMM}{The created DMM}
  \item{DMM_bin}{The DMM in a binary version}
  \item{SDC}{The created SDC.}
  \item{SDC_err}{Percentage error between input SDC and created SDC.}
}
}
\description{
Creates a \code{N_src} x \code{N_tgt}  domain mapping matrix (DMM) by reaching a certain given SDC.
}
\examples{
set.seed(1234)

crt_DMM(N_src=7,
        N_tgt=16,
        DMM_PAR=0.4,
        upper_Bound=1)

}
