% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_DOMAINS.R
\name{crt_DOMAINS}
\alias{crt_DOMAINS}
\title{Create EAD Domains}
\usage{
crt_DOMAINS(
  P_FD,
  N_DD,
  N_PrD,
  N_RD,
  DMM_PAR,
  DSM_method = c("DNS", "modular"),
  DSM_param,
  ut_DMM = F,
  uB_DMM = 1,
  ub_DSM = 1,
  allowZero = F
)
}
\arguments{
\item{P_FD}{The product matrix representing the functional requirements (FRs)}

\item{N_DD}{The number of elements in the physical domain (PD)}

\item{N_PrD}{The number of elements in the process domain (PrD)}

\item{N_RD}{The number of elements in the resource domain (RD)}

\item{DMM_PAR}{System design complexity for the DMMs. In the current version this values is applied for all domains.}

\item{DSM_method}{A method how to generate the DSM pattern. The following options are available:
\describe{
  \item{DNS}{The DSM is generated by treating the \code{DSM_param} input as a desired density.}
  \item{modular}{See \link[EAD]{crt_DSM} for details.}
}}

\item{DSM_param}{Either the adjusted and normalized structural complexity measure or bounds \code{c(UB,LB)} for the desired density of DSM. The input depends on the selected \code{DSM_method}.
In the current version this values is applied for all domains.}

\item{ut_DMM}{Boolean. If \code{ut_DMM=TRUE}, the upper triangle DMMs are created representing indirect inter domain dependencies.}

\item{uB_DMM}{The upper bound for integer sampling with the DMM matrices. Default \code{uB_DMM=1} which means that only binary DMMs are created.}

\item{ub_DSM}{The upper bound for integer sampling with the DSM matrices. Default \code{uB_DSM=1} which means that only binary DSMs are created.}

\item{allowZero}{Boolean. If \code{allowZero=TRUE} the DMM can contain empty columns. However, the procedure ensures that there are no unused domain elements. Therefore, specific DSM entries are forced to be none-zero.
For further detail see the \code{forced} parameter in \link[EAD]{crt_DSM}.}
}
\value{
Returns an EAD object containing
\describe{
  \item{P}{A list of product matrices starting with the product represent in the functional domain (FD), the physical domain (PD), the proccess domain (PrD) and the resource domain (RD)}
  \item{DSM}{A list containing the Design Structure Matrices (DSM) for each domain.}
  \item{DMM}{A list containing the Domain Mapping Matrices (DMM) for each domain.}
  \item{measures}{A named list containing measures on system and product level.}
  \item{message}{A string containing messages from the simulation routine.}
}
}
\description{
This function creates EAD domains and returns the design structure matrices (DSMs), the domain mapping matrices (DMMs) and the product matrices (P).
The following matrices are created
}
\examples{
require(tidyr)
P_FD<-expand_grid(FR1=c(0,1),
FR2=c(0,1),
FR3=c(0,1),
FR4=c(0,1))

## creates a EAD by using the Density method
## In each DSM 10\% of the entries are none zero
EAD<-crt_DOMAINS(P_FD=P_FD,
            N_DD=15,
            N_PrD=30,
            N_RD=50,
            DMM_PAR=c(0,1),
            DSM_method="DNS",
            DSM_param=c(0,0.2),
            ut_DMM=F)

EAD$P$RD # for RES_CONS_PAT
EAD$DSM$DD # for the DSM in the physical domain
EAD$measures for the measures
}
