% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_DSM.R
\name{crt_DSM}
\alias{crt_DSM}
\title{Create a domain dependency matrix (DSM)}
\usage{
crt_DSM(N_el, SC_adj_n_in = NULL, DENS = NULL, upper_Bound = 5, forced = NULL)
}
\arguments{
\item{N_el}{Number of rows and columns for the DSM.}

\item{SC_adj_n_in}{Double with \code{0<=SC_adj_n_in<1}. Specifies the system design complexity which should be generated. For further details see \link[EAD]{measure_structuralcomplexity}.}

\item{DENS}{A numeric value specifying the density within the DSM. This is an alternative method to generate the DSM.}

\item{upper_Bound}{Integer, default \code{upper_Bound=20}, specifies the maximum number in the DMM if \code{binary=FALSE}}

\item{forced}{A numeric vector containing indices of elements which cannot be zero since otherwise a invalid design is created later on.}
}
\value{
A list object containing the product mix. This list consists of:
\describe{
  \item{DSM}{The created DSM}
  \item{DENS_DSM}{Density of the DSM}
  \item{SC_adj_n}{The achieved adjusted and normalized structural complexity measure.}
}
}
\description{
Creates a \code{N_el} x \code{N_el}  domain dependency matrix (DSM) by trying to reach a given SC_adj_n or a DENS value.
}
\examples{
set.seed(1234)

crt_DSM(N_el=7,SC_adj_n_in=0.01,upper_Bound=5)
}
