% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_EAD.R
\name{crt_EAD_MC}
\alias{crt_EAD_MC}
\title{Create EADs using a design of experiments (DOE)}
\usage{
crt_EAD_MC(
  DOE,
  NUMB_CORES = 4,
  logfile = "",
  time_limit = Inf,
  cl = NULL,
  ehNodes = "remove"
)
}
\arguments{
\item{DOE}{A data.frame, where each row contain a parameter set for an EAD realization. The data frame contain the following columns.
\describe{
  \item{N_FR}{The number of elements in the functional domain (FD)}
  \item{N_DD}{The number of elements in the physcial domain (PD)}
  \item{N_PrD}{The number of elements in the process domain (PrD)}
  \item{N_RD}{The number of elements in the resource domain (RD)}
  \item{DNS}{Density of the product mix \code{P_FD}.}
  \item{N_PROD}{Number of products (rows) in \code{P_FD}.}
  \item{method_FD}{Method for generating the product mix. For available methods see: \link[EAD]{create_ProductMix}.}
  \item{TOTAL_DEMAND}{Total demand}
  \item{Q_VAR}{Demand heterogeneity.}
  \item{DMM_PAR}{Parameter for creating the DMM matrices. Named list for each of the three DMMs.
  The input is list object, where each list element is a 2d-vector specifying the lower and upper bound of the desired standardized system design complextiy.
  For further details see: \link[EAD]{crt_DMM}.}
  \item{uB_DMM}{The upper bound for entries in the DMM matrices. Passed to: \link[EAD]{crt_DMM}.}
  \item{allowZero}{Specifies whether zero columns are allowed in the DMM matrices. Passed to: \link[EAD]{crt_DMM}.}
  \item{ut_DMM}{Boolean. If \code{ut_DMM=TRUE}, the upper triangle DMMs (DMM_FD_PrD,DMM_FD_RD,DMM_PD_RD) are created representing indirect inter domain dependencies.}
  \item{DSM_param}{The first two entries refer to the density of the DSMs and the second pair to the cv if the DSM_method='modular' is used. For details see: \link[EAD]{crt_DSM}.}
  \item{DSM_method}{The method for DSM creation. For details see: \link[EAD]{crt_DSM}.}
  \item{ub_DSM}{Upper bounds for entries in the DSMs.}
  \item{TC}{Total costs}
  \item{r_in}{Proportion of indirect costs on total costs}
  \item{r_fix}{Proportion of fixed costs on total indirect costs}
  \item{RC_cor}{Correlation between variable cost vector and fixed cost vector}
  \item{RC_cv}{Coefficient of variation for resource cost distribution}
  \item{N_RUN}{Number of runs per unique setting}
}}

\item{NUMB_CORES}{Number of cores to run the simulation. Default: \code{NUMB_CORES=4}}

\item{logfile}{Optional argument, specifying the name of the logfile which is created in the working directory.}

\item{time_limit}{Time in seconds after which each EAD creation is interrupted. This is necessary since there are certain input combinations which result in long computational time.
Therefore runs longer than \code{time_limit} are excluded from the output. The default value is infinity which means that no interruption is done.}

\item{cl}{An optional cluster object created by \link[parallel]{makeCluster}. This allows to distribute the computation across different nodes within a network.
For further details see the documentation of the 'parallel' package. By default the function \link[EAD]{setupMC} is called.}

\item{ehNodes}{A character specifying the parallel node behavior on errors. This argument is passed to the '.errorhandling' of \link[foreach]{foreach}. The default value is \code{'remove'}.}
}
\value{
A list of EAD objects
}
\description{
This function creates EAD realizations using a design of experiments (DOE). The design is specified by \code{DOE}
}
\examples{
set.seed(1234)

library(tidyr)
DOE<-expand_grid(N_FR = list(c(15)),
                 N_DD = list(c(18,26)),
                 N_PrD = list(c(36,52)),
                 N_RD = list(c(72,104)),
                 DNS = list(c(0.07,0.5)),
                 N_PROD = list(c(50)),
                 method_FD = "DNS",
                 TOTAL_DEMAND = list(c(1000,1000)),
                 Q_VAR = list(c(0,3)),
                 DMM_PAR = expand_grid(FD_PD = list(c(0,0.08)),
                                       PD_PrD = list(c(0,0.1)),
                                       PrD_RD = list(c(0,0.1))),
                 uB_DMM = 1,
                 allowZero = F,
                 ut_DMM = F,
                 DSM_param=expand_grid(PD = list(c(0,0.05,0,1)),
                                       PrD = list(c(0,0.05,0,1)),
                                       RD = list(c(0,0.05,0,1))), # first two entries refer to the density of the DSMs and the second pair to the cv if the DSM_method='modular' is used.
                 DSM_method='modular',
                 ub_DSM = 1,
                 TC = 10^6,
                 r_in = list(c(0,0.9)),
                 r_fix = list(c(0,1)),
                 RC_sdlog = list(c(0,1.5)),
                 N_RUN = 1:4 # number of runs
)

EAD <- crt_EAD_MC(DOE,NUMB_CORES=4,logfile="log.txt")
EAD[[1]]

}
