% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DANI_DMM}
\alias{DANI_DMM}
\title{A DMM containing the dependencies dependencies between functional requirements and design parameter as reported by \insertCite{Danilovic.2007;textual}{EAD}.}
\format{
A DMM with 18 by 41 elements.
}
\usage{
data(DANI_DMM)
}
\description{
A DMM containing the dependencies dependencies between functional requirements and design parameter as reported by \insertCite{Danilovic.2007;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
