% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Tipologia}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19145 rows and 19 columns.
}
\usage{
data
}
\description{
Banco de dados da tipologia
}
\keyword{datasets}
