% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_Driver.R
\name{development_costs}
\alias{development_costs}
\title{Calculates Components' Development and Part Administration Costs}
\usage{
development_costs(P_PD, DEMAND, PDC_fix, R_dvl, R_PA)
}
\arguments{
\item{P_PD}{The product mix in the physical domain containing products in rows and components in columns.}

\item{DEMAND}{A numeric demand vector. If products are excluded from the product mix, the entries are zero.}

\item{PDC_fix}{A numeric vector holding the fixed costs for a component. These costs are calculated by tracing resource costs back into the physical domain.}

\item{R_dvl}{Proportion of development costs on \code{PDC_fix}.}

\item{R_PA}{Proportion of part administration costs on \code{PDC_fix}.}
}
\value{
Returns the summed development costs and part administration costs
}
\description{
This function calculates components' development and part administration costs.
}
\details{
For further details see \insertCite{Meerschmidt.2024;textual}{EAD}
}
\examples{

require(EAD)
EAD <- smallEAD
P_PD <- EAD$P$PD
DEMAND<-c(10,50,100,30)
PDC_fix<-c(700,1819,4589,1367)
R_dvl<-0.3
R_PA<-0.2

development_costs(P_PD,
                  DEMAND,
                  PDC_fix,
                  R_dvl,
                  R_PA)
}
\references{
\insertAllCited{}
}
