% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ead-package.R
\docType{package}
\name{ead-package}
\alias{ead}
\alias{ead-package}
\title{ead: Episódios de autocratização e democratização}
\description{
ead codifica episódios de mudança de regime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arturhgq/ead}
  \item \url{https://arturhgq.github.io/ead/}
  \item Report bugs at \url{https://github.com/arturhgq/ead/issues}
}

}
\author{
\strong{Maintainer}: Artur Quirino \email{contact@arturhgq.com} (\href{https://orcid.org/0000-0001-8071-6348}{ORCID})

}
\keyword{internal}
