% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{ead_messages}
\alias{ead_messages}
\title{Imprime no console os parâmetros da tipologia}
\usage{
ead_messages(.dem_parameters, .aut_parameters, .index_string)
}
\arguments{
\item{.dem_parameters}{Veja \code{\link[=typology]{typology()}}}

\item{.aut_parameters}{Veja \code{\link[=typology]{typology()}}}

\item{.index_string}{Índice de democracia}
}
\description{
Imprime no console os parâmetros da tipologia
}
