% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{epi_coding}
\alias{epi_coding}
\title{Codifica os episódios de mudança de regime}
\usage{
epi_coding(.data, .measure, .countries, .group_id)
}
\arguments{
\item{.data}{dados do projeto \emph{"Varieties of Democracy"}.}

\item{.measure}{"autocratization" ou "democratization"}

\item{.countries}{países}

\item{.group_id}{agrupamentos dos episódios de mudança de regime}
}
\description{
Codifica os episódios de mudança de regime
}
