% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{epi_id}
\alias{epi_id}
\title{Cria um identificar único para cada episódio de mudança de regime}
\usage{
epi_id(.data, .measure, .countries, .group_id)
}
\arguments{
\item{.data}{dados do projeto \emph{"Varieties of Democracy"}.}

\item{.measure}{"autocratization" ou "democratization"}

\item{.countries}{países}

\item{.group_id}{agrupamentos dos episódios de mudança de regime}
}
\description{
Cria um identificar único para cada episódio de mudança de regime.
}
