% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{extend_group}
\alias{extend_group}
\title{Estende os agrupamentos prévios.}
\usage{
extend_group(
  .data,
  .countries,
  .define_epi,
  .threshold,
  .before_value,
  .after_value,
  .before_value_interval_expr,
  .after_value_interval_expr
)
}
\arguments{
\item{.data}{dados do projeto \emph{"Varieties of Democracy"}.}

\item{.countries}{países}

\item{.define_epi}{Lista de parâmetros que delimitam a extensão dos episódios
de mudança de regime.
\itemize{
\item \code{.stop}: intervalo temporal que encerra os episódios de
mudança de regime (em anos).
}

Por definição, quando a variação no índice de democracia for < 0.01
por três anos consecutivos em casos de democratização, e > -0.01 em casos
de autocratização, \code{.stop} encerra os episódios de mudança de
regime. Durante os episódios, variações menores que o \code{.threshold}
indicam: 1) eventos de reação às alterações no status quo ou 2) a escasez de
recursos para a implementação de novas alterações no regime.
\itemize{
\item \code{.keep}: número de eventos de estagnação ou reação aceitos
durante episódios de mudança. Por definição, dois eventos.
\item \code{.cumulative}: define o valor acumulado mínimo para codificar
um episódio de mudança de regime.
}

Por definição, 10\% (0,1) em democratizações e -10\% (-0,1) em autocratizações.
\itemize{
\item \code{.after}: quantidade máxima de anos em que o índice de democracia
poderá não atingir o ponto de corte e ainda assim ter esses anos incluídos
no calculo dos episódios de mudança de regime.
Depende de \code{.after_value}.
\item \code{.before}: quantidade máxima de anos em que o índice de democracia
poderá não atingir o ponto de corte, mas ainda assim ter esses anos incluídos
no calculo dos episódios de mudança de regime.
Depende de \code{.before_value}.
\item \code{.after_value}: variação mínima, inferior ao \code{.threshold},
para extensão do período de encerramento dos episódios.
\item \code{.before_value}: variação mínima, inferior ao \code{.threshold},
para extensão do período de abertura dos episódios.
}}

\item{.threshold}{Define o valor mínimo de variação no
índice de democracia para determinar o ínicio de processos de mudança de regime.
\code{.threshold} aceita valores entre 0 e 1.

Por definição, 0.01(1\%) em episódios de democratização e -0.01(-1\%) em
episódios de autocratização.}

\item{.before_value}{Veja \code{\link[=typology]{typology()}}}

\item{.after_value}{Veja \code{\link[=typology]{typology()}}}

\item{.before_value_interval_expr}{expressão definida em \code{measures_exprs()}
que estima a variação no índice de democracia, inferior ao \code{.threshold},
usada para alterar o ano de abertura dos episódios de mudança de regime.
dada}

\item{.after_value_interval_expr}{expressão definida em \code{measures_exprs()}
que estima a variação mínima no índice de democracia, inferior ao \code{.threshold},
usada para alterar o ano de encerramento dos episódios de mudança de regime.}
}
\description{
Redesenha o início e fim dos episódios de acordo com
os parâmetros .after, .before, .after_value e .before_value.
O objetivo é ampliar a quantidade de casos cobertos.
}
