% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{final_group}
\alias{final_group}
\title{Agrupamento final}
\usage{
final_group(.data, .countries)
}
\arguments{
\item{.data}{dados do projeto \emph{"Varieties of Democracy"}.}

\item{.countries}{países}
}
\description{
Agrupamento final
}
\note{
Cada grupo corresponde a um episódio de mudança de regime
}
