% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gearbox_DMM_PD_PrD}
\alias{gearbox_DMM_PD_PrD}
\title{A DMM containing the component to process mapping of gearbox used by \insertCite{Bonjour.2010;textual}{EAD}.}
\format{
A DMM with 12 by 21 elements.
}
\usage{
data(gearbox_DMM_PD_PrD)
}
\description{
A DMM containing the component to process mapping of gearbox used by \insertCite{Bonjour.2010;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
