% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gearbox_DMM_PrD_RD}
\alias{gearbox_DMM_PrD_RD}
\title{A DMM containing the process to resource mapping of gearbox used by \insertCite{Bonjour.2010;textual}{EAD}.}
\format{
A DMM with 21 by 14 elements.
}
\usage{
data(gearbox_DMM_PrD_RD)
}
\description{
A DMM containing the process to resource mapping of gearbox used by \insertCite{Bonjour.2010;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
