% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gearbox_DSM_PrD}
\alias{gearbox_DSM_PrD}
\title{A DSM containing the process to process dependencies of gearbox used by \insertCite{Bonjour.2010;textual}{EAD}.}
\format{
A DSM with 21 by 21 elements.
}
\usage{
data(gearbox_DSM_PrD)
}
\description{
A DSM containing the process to process dependencies of gearbox used by \insertCite{Bonjour.2010;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
