% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gearbox_DSM_RD}
\alias{gearbox_DSM_RD}
\title{A DSM containing the resource to resource dependencies of gearbox used by \insertCite{Bonjour.2010;textual}{EAD}.}
\format{
A DSM with 14 by 14 elements.
}
\usage{
data(gearbox_DSM_RD)
}
\description{
A DSM containing the resource to resource dependencies of gearbox used by \insertCite{Bonjour.2010;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
