% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFirm.R
\name{genFirm}
\alias{genFirm}
\title{Generates a firm}
\arguments{
\item{ProductionEnvironment}{A structured list. See 'examples' for further details.  \cr}

\item{DISP1}{The number of top rescources which should allocate DISP2 percent of total costs.
Therefore 1<=DISP1<NUMB_RES.}

\item{DISP2}{list containing the elements \code{DISP2$lb} for the lower and \code{DISP2$ub} for the upper bound of DISP2.
The function chooses DISP2 from a uniform distribution. If lb=ub the exact value is choosen.
DISP2 specifies how mutch percentage of total costs are assigned to the top DISP1 resources? DISP2 0<DISP2<1 \cr}

\item{DENS}{The density (sparsity) of the RES_CONS_PAT. A list containing the elements \code{DENS$lb} for the lower and \code{DENS$ub} for the upper bound of DENS.}

\item{COR1}{A list containing the elements \code{COR1$lb} for the lower and \code{COR1$ub} for the upper bound of COR1.}

\item{exogenious}{Default=TRUE. If set to true the product mix decision is exognious otherwise endogenious.}
}
\description{
Generates a firm by Anand et al. (2017) procedure.
}
\examples{
ProductionEnvironment<-list()
ProductionEnvironment[['NUMB_RES']]<-50 #Amount of resources
ProductionEnvironment[['NUMB_PRO']]<-50 #Amount of products
ProductionEnvironment[['TC']]<-10^6 # Total


DISP1<-10
DISP2<-list(lb=0.3,ub=0.3)
DENS<-list(lb=0.2,ub=0.2)
COR1<-list(lb=0.2,ub=0.2)
COR2<-list(lb=0.2,ub=0.2)

genFirm(ProductionEnvironment,DISP1,DISP2,DENS,COR1,COR2,exogenious=TRUE)
}
