% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRES_CONS_PAT.R
\name{genRES_CONS_PAT}
\alias{genRES_CONS_PAT}
\title{Generates a Resorce Consumption Pattern Matrix}
\usage{
genRES_CONS_PAT(ProductionEnvironment, DENS, COR1, COR2, RES_METHOD = "ANAND")
}
\arguments{
\item{ProductionEnvironment}{A structured list. See 'examples' for further details.  \cr}

\item{DENS}{number of customers which are generated.}

\item{COR1}{Correlation on how the unit size resources correlate with the base. Can be a list in the format list(lb=...,ub=...),
where lb stands for the lower bound of COR1 and ub for upper bound. See Anand et al. (2017) for further details.}

\item{COR2}{Correlation value of how the batch resources should correlate. Base on the choosen method the base is different.
See 'RES_method' for further details."}

\item{RES_METHOD}{Default: "ANAND". Options are: "KGM" and "BALA". The original implementation
from Anand et al. (2017) generates correlated numbers where COR2 ist the correlation between base and batch resources.
"KGM" and  "BALA" generates the batch resource correlation based on the unit size. THis is equivalent to Balakrishnan et al. (2011)}
}
\value{
Returns RES_CONS_PAT and some statistics as list. \cr
}
\description{
Based on the choosen method a RES_CONS_PAT is created by this function.
}
\examples{
ProductionEnvironment[['NUMB_RES']]<-NUMB_RES #Amount of resources
ProductionEnvironment[['NUMB_PRO']]<-NUMB_PRO #Amount of product

ProductionEnvironment[['UnitSize']]<-floor(VOL_SHARE_RES*ProductionEnvironment[['NUMB_RES']])
ProductionEnvironment[['BatchSize']]<-floor((1-VOL_SHARE_RES)*ProductionEnvironment[['NUMB_RES']])

# THe base and unit resources correlate with the coefficient of 0.8
# unit size resources and batch res. correlate with a level of 0.33 cause of choosen method
genRES_CONS_PAT(ProductionEnvironment,DENS=0.4,COR1=0.8,COR2=0.33,RES_METHOD="BALA")

}
