% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{glue_coding}
\alias{glue_coding}
\title{Cola os episódios de mudança de regime em uma única coluna}
\usage{
glue_coding(.data, .measure, .countries, .group_id)
}
\arguments{
\item{.data}{dados do projeto \emph{"Varieties of Democracy"}.}

\item{.measure}{"autocratization" ou "democratization"}

\item{.countries}{países}

\item{.group_id}{agrupamentos dos episódios de mudança de regime}
}
\description{
Cola os episódios de mudança de regime em uma única coluna
}
