% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{group_end}
\alias{group_end}
\title{Identifica o final do agrupamentos}
\usage{
group_end(.data)
}
\arguments{
\item{.data}{\code{ead_measures} data frame}
}
\description{
Identifica o final do agrupamentos
}
\note{
Cada agrupamento corresponde a um episódio de mudança de regime.
}
