% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{group_start}
\alias{group_start}
\title{Identifica o início dos agrupamentos}
\usage{
group_start(.data)
}
\arguments{
\item{.data}{\code{ead_measures} data frame}
}
\description{
Identifica o início dos agrupamentos
}
\note{
Cada agrupamento corresponde a um episódio de mudança de regime.
}
