% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LI_valve}
\alias{LI_valve}
\title{A DMM containing the dependencies between functional requirements and design parameter for a valve as reported by \insertCite{Li.2008;textual}{EAD}.}
\format{
A DMM with 29 by 49 elements.
}
\usage{
data(LI_valve)
}
\description{
A DMM containing the dependencies between functional requirements and design parameter for a valve as reported by \insertCite{Li.2008;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
