% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_CI}
\alias{measure_CI}
\title{Calculates the Commonality Index (CI)}
\usage{
measure_CI(P)
}
\arguments{
\item{P}{A product matrix containing the products in rows and and features, components, process or resources in columns.}
}
\value{
Returns the CI value
}
\description{
This function calculates the commonality index as defined by \insertCite{MartinIshii1996;textual}{EAD}
}
\examples{

P<-matrix(c(1,0,0,1,
              0,1,0,0,
              1,0,0,0,
              0,1,1,0,
              0,1,0,1),
       nrow = 4,
       ncol = 5)
measure_CI(P)
}
\references{
\insertAllCited{}
}
