% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_CoupComp}
\alias{measure_CoupComp}
\title{Calculates the Coupling Complexity (CC)}
\usage{
measure_CoupComp(DSM)
}
\arguments{
\item{DSM}{A dependency structure matrix, reflecting the interfaces between elements.}
}
\value{
Returns the CC value
}
\description{
This function coupling complexity (CC) as defined by \insertCite{Ameri.2008;textual}{EAD} and \insertCite{Summers.2010;textual}{EAD}.
Note, this function is not implemented yet since description of authors is not sufficient to reproduce the algorithm.
}
\examples{
require(EAD)

## water Sprinkler example by Ameri et al. (2008) ##
data("waterSprinkler")
DSM<-waterSprinkler
measure_CoupComp(DSM)
}
\references{
\insertAllCited{}
}
