% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_DENS}
\alias{measure_DENS}
\title{Measure density of a product matrix}
\usage{
measure_DENS(P)
}
\arguments{
\item{P}{A product matrix (e.g, RES_CONS_PAT) with products in rows and features, components, process or resources in columns.}
}
\value{
Returns the corresponding density value
}
\description{
This function measures the density defined as the proportion of non-zero entries in P.
This measure is used in several studies \insertCite{Anand2019,BLH2011}{EAD}.
}
\examples{

data('csd_EAD')
P<-CSD_EAD$RES_CONS_PAT

measure_DENS(P)
}
\references{
\insertAllCited{}
}
