% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_designComplexity}
\alias{measure_designComplexity}
\title{Calculate Design Complexity}
\usage{
measure_designComplexity(A, norm = T)
}
\arguments{
\item{A}{A domain mapping matrix (DMM) or dependency structure matrix (DSM), reflecting the transition between two domains.}

\item{norm}{Boolean, default\code{TRUE}. Decides if the normalized design complexity measures should be returned.}
}
\value{
The calculated design complexity measure.
}
\description{
This measure is based on the Shannon entropy. Adapted from \insertCite{Modrak.2015;textual}{EAD}.
}
\examples{

A<-matrix(c(1,0,0,
            1,1,0,
            1,1,1),
         nrow=3,
         ncol=3,
         byrow=T)

measure_designComplexity(A)

### Example of Modrak & Bednar (2015)  - Figure 4
DMM_a<-matrix(c(1,0,
1,1),byrow = T,nrow=2)
measure_designComplexity(DMM_a,norm=F) # should be 1.39


DMM_b<-matrix(c(1,0,0,0,
                1,1,0,0,
                1,0,1,0,
                1,0,0,1,
                1,1,1,0,
                1,1,0,1,
                1,0,1,1,
                1,1,1,1),byrow = T,nrow=8)
measure_designComplexity(DMM_b,norm=F) # should be 33.3

}
\references{
\insertAllCited{}
}
