% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_DISS}
\alias{measure_DISS}
\title{Calculate normalized dissimilarity}
\usage{
measure_DISS(P, norm = T)
}
\arguments{
\item{P}{A product matrix with products in rows and attributes in columns.}
}
\value{
The calculated normalized dissimilarity measure.
}
\description{
Takes a product matrix with products in rows and attributes in columns. This measure is defined by Buchholz (2012).
}
\examples{

P<-matrix(c(1,1,1,1,
            0,1,1,0,
            0,0,1,1),
           nrow = 4,
           ncol = 3)

measure_DISSnrom(P)
}
