% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_DISTcv}
\alias{measure_DISTcv}
\title{product distance mean coefficient of variation.}
\usage{
measure_DISTcv(P)
}
\arguments{
\item{P}{A product matrix with products in rows and attributes in columns.}
}
\value{
The coefficient of variation of products' mean distances.
}
\description{
Calculates the mean distance of each product. This vector is used to calculate the coefficient of variation (cv) in a second step.
A cv of zero means, that all products are equally distributed. Larger cv values imply a more heterogeneous product mix.
}
\examples{

P<-matrix(c(1,1,1,1,0,
            0,1,1,0,1,
            0,0,1,1,1),
           nrow = 5,
           ncol = 3)

measure_DISTcv(P)
}
