% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_DISTsum}
\alias{measure_DISTsum}
\title{Calculate sum of distances}
\usage{
measure_DISTsum(P)
}
\arguments{
\item{P}{A product matrix with products in rows and attributes in columns.}
}
\value{
The calculated sum of distances.
}
\description{
Calculates the absolute sum of euclidean distances for a given matrix with elements in rows and attributes in columns.
For further details on this measure see \insertCite{Buchholz.2012;textual}{EAD}.
}
\examples{

P<-matrix(c(1,1,1,1,
            0,1,1,0,
            0,0,1,1),
           nrow = 4,
           ncol = 3)

measure_DISTsum(P)
}
\references{
\insertAllCited{}
}
