% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_diversificationINDEX}
\alias{measure_diversificationINDEX}
\title{Calculate the Product Diversification Index}
\usage{
measure_diversificationINDEX(P, DMD = NULL)
}
\arguments{
\item{P}{A product matrix (e.g, RES_CONS_PAT) with products in rows and features, components, process or resources in columns.}

\item{DMD}{A demand vector. If no demand vector is given, then a normal distribution is assumed where each product has the quantity one.}
}
\value{
Returns the corresponding index value
}
\description{
Calculates the Herfindahl-based product diversification index (D) as defined by \insertCite{Gollop.1997;textual}{EAD} and \insertCite{Gollop.1991;textual}{EAD}.
}
\examples{

data('csd_EAD')
P<-CSD_EAD$RES_CONS_PAT
DMD<-crt_DEMAND(NROW(P),1000,1)$DEMAND

measure_diversificationINDEX(P,DMD)
}
\references{
\insertAllCited{}
}
