% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_GCI}
\alias{measure_GCI}
\title{Calculates the Generalized Complexity Index (GCI)}
\usage{
measure_GCI(P)
}
\arguments{
\item{P}{A product matrix containing the products in rows and and features, components, process or resources in columns.}
}
\value{
Returns the GCI value
}
\description{
This function calculates the generalized complexity index (GCI) as defined by \insertCite{JACOBS2013;textual}{EAD}
}
\examples{

P<-matrix(c(1,0,0,1,
              0,1,0,0,
              1,0,0,0,
              0,1,1,0,
              0,1,0,1),
       nrow = 4,
       ncol = 5)
measure_GCI(P)
}
\references{
\insertAllCited{}
}
