% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_HVM}
\alias{measure_HVM}
\title{Calculates the Halstead-derived volume measure complexity (HVM)}
\usage{
measure_HVM(DSM, norm = T)
}
\arguments{
\item{DSM}{A dependency structure matrix, reflecting the interfaces between elements.}

\item{norm}{Boolean, default=T normalizes the HVM in a range of 0-1 where HVM=1 is the maximum possible value for a given matrix.}
}
\value{
Returns the HVM value
}
\description{
This function calculates the Halstead-derived volume measure complexity (HVM) as defined by \insertCite{Halstead.1979;textual}{EAD}, \insertCite{Prather.1984;textual}{EAD} and \insertCite{Hennig.2022;textual}{EAD}.
}
\examples{
require(EAD)

## water Sprinkler example by Ameri et al. (2008) ##
data("waterSprinkler")
DSM<-waterSprinkler

measure_HVM(DSM)
}
\references{
\insertAllCited{}
}
