% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_INTER}
\alias{measure_INTER}
\title{Calculate inter-product heterogeneity}
\usage{
measure_INTER(P)
}
\arguments{
\item{P}{A product matrix (e.g, RES_CONS_PAT) with products in rows and features, components, process or resources in columns.}
}
\value{
Returns the calculated inter value.
}
\description{
Calculates the INTRA measure defined by Gupta (1993) and adapted by Mertens (2020).
}
\examples{

P<-matrix(c(2,1,1,1,
           2,1,1,3,
           2,2,2,2,
           1,1,1,1),
         nrow = 4,
         ncol = 4,byrow=T)
measure_INTER(P)
}
