% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_INTRA}
\alias{measure_INTRA}
\title{Calculate intra-product heterogeneity}
\usage{
measure_INTRA(P, norm = F)
}
\arguments{
\item{P}{A product matrix (e.g, RES_CONS_PAT) with products in rows and features, components, process or resources in columns.}

\item{norm}{Boolean, default \code{FALSE}. If set to \code{TRUE}, then the normalized measures is used.}
}
\value{
Returns the calculated intra value.
}
\description{
Calculates the INTRA measure defined by \insertCite{Gupta.1993;textual}{EAD} and adapted by \insertCite{Mertens.2020;textual}{EAD}.
}
\examples{

P<-matrix(c(2,1,1,1,
           2,1,1,3,
           2,2,2,2,
           1,1,1,1),
         nrow = 4,
         ncol = 4,byrow=T)
measure_INTRA(P)
}
\references{
\insertAllCited{}
}
