% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_JSDC}
\alias{measure_JSDC}
\title{Calculates Jung's system design complexity}
\usage{
measure_JSDC(DMM, norm = F)
}
\arguments{
\item{DMM}{A domain mapping matrix, reflecting the transition between two domains.}
}
\value{
The calculated system design complexity measure.
}
\description{
This measure calculates the system design complexity  for a given domain mapping matrix (DMM) according to \insertCite{Jung.2022;textual}{EAD}.
}
\examples{
## example as shown in Jung et al. (2022)
data('RDU1')
measure_JSDC(RDU1) # must be 389.5


data('RDU2')
measure_JSDC(RDU2) # must be 466.1
}
\references{
\insertAllCited{}
}
