% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_LOF}
\alias{measure_LOF}
\title{Calculate Local Outlier Factor (LOF)}
\usage{
measure_LOF(P, perc = 0.1)
}
\arguments{
\item{P}{A product matrix (e.g, RES_CONS_PAT) with products in rows and features, components, process or resources in columns.}

\item{n}{A value between \code{2<n<NROW(P)} indicating the number of products used to calculate the LOF's.}
}
\value{
Returns the corresponding index value
}
\description{
Calculates the local outlier factor of observation (rows) in \code{P} as defined by \insertCite{Breunig2000;textual}{EAD} and \insertCite{XU2022;textual}{EAD} The input is passed to \link[DescTools]{LOF}
}
\examples{

data('csd_EAD')
P<-CSD_EAD$RES_CONS_PAT

measure_LOF(P,n=2)
}
\references{
\insertAllCited{}
}
