% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_modularity}
\alias{measure_modularity}
\title{measure_modularity}
\usage{
measure_modularity(A, preMember = NULL, plot_op = F)
}
\arguments{
\item{A}{A data.frame containing the dependency structure matrix with information on element dependencies.}

\item{preMember}{A numerical vector containing the assignment of vertices to cluster.
The default value is \code{preMember=NULL} which means that the communities are identified using the \link[igraph]{fastgreedy.community}}

\item{plot_op}{Boolean, default \code{plot_op=FALSE}. If set to true the graph is visualized.}
}
\value{
The calculated modularity measure.
}
\description{
Calculates the modularity measure of a undirected network as described by \insertCite{Blondel.2008;textual}{EAD}, \insertCite{Newman.2006;textual}{EAD} and \insertCite{Leicht.2008;textual}{EAD}.
The implementation was checked by using the train boogie data set as reported by \insertCite{Sinha.2018;textual}{EAD}.
}
\examples{

## unsymmetrical matrix as input
## is transformed into a symmetric matrix first
## use the undirected approach by:
   ### Blondel, Guillaume, Lambiotte, & Lefebvre (2008)
    ### well as Newman (2006)
  DSM<-matrix(c(1,0,0,
              1,1,0,
              1,1,1),
           nrow=3,
           ncol=3,
           byrow=T)
  measure_modularity(DSM,symmetric=T)

### Example presented by Sinha & Suh (2018)
  library(EAD)
  DSM<-trainbogie
  modules<-sapply(strsplit(rownames(DSM),"_"),function(x) x[1])

  res<-measure_modularity(DSM,
                          preMember = as.numeric(factor(modules))
  res # should be 0.74

}
\references{
\insertAllCited{}
}
