% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_neumannEntropy}
\alias{measure_neumannEntropy}
\title{Calculate the von Neumann entropy for graphs.}
\usage{
measure_neumannEntropy(A, norm = F)
}
\arguments{
\item{DSM}{A dependency structure matrix, reflecting the interfaces between elements.}
}
\value{
The calculated von Neumann entropy measure.
}
\description{
This function takes a DSM and calculates the von Neumann entropy. For further details see: \insertCite{Passerini.2009;textual}{EAD}.
}
\examples{

DSM<-matrix(c(1,0,0,
            1,1,0,
            1,1,1),
         nrow=3,
         ncol=3,
         byrow=T)

measure_neumannEntropy(DSM)
}
\references{
\insertAllCited{}
}
