% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_NPV}
\alias{measure_NPV}
\title{Measure the proportion of product variants}
\usage{
measure_NPV(P)
}
\arguments{
\item{P}{A product matrix (e.g, RES_CONS_PAT) with products in rows and features, components, process or resources in columns.}
}
\value{
Returns the NPV value
}
\description{
This function measures the proportion of product variants present in \code{P} compared to the free combination.
}
\examples{

P<-matrix(c(1,0,0,1,
              0,1,0,0,
              1,0,0,0,
              0,1,1,0,
              0,1,0,1),
       nrow = 4,
       ncol = 5)
measure_NPV(P)
}
