% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_PCI}
\alias{measure_PCI}
\title{Calculate the Product Line Commonality Index}
\usage{
measure_PCI(P)
}
\arguments{
\item{P}{A product matrix representing the components per product (P_DD) with products in rows and components in columns.}
}
\value{
Returns the corresponding index value
}
\description{
This index was introduced by Kota et al. (2000). The f-values as defined by Kota are set to one.
}
\examples{

P<-matrix(c(1,0,0,1,
              0,1,0,0,
              1,0,0,0,
              0,1,1,0,
              0,1,0,1),
       nrow = 4,
       ncol = 5)

measure_PCI(P)
}
