% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_semiangularity}
\alias{measure_semiangularity}
\title{Calculate Semiangularity}
\usage{
measure_semiangularity(A)
}
\arguments{
\item{DMM}{A domain mapping matrix, reflecting the transition between two domains.}
}
\value{
The calculated semiangularity measure.
}
\description{
This measure calculates the semiangularity for a given domain mapping matrix (DMM) according to Suh (1995) and Delaš et al. (2018).
}
\examples{

DMM<-matrix(c(1,0,0,
            1,1,0,
            1,1,1),
         nrow=3,
         ncol=3,
         byrow=T)

measure_semiangularity(DMM)
}
