% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_structuralcomplexity}
\alias{measure_structuralcomplexity}
\title{Calculate structural complexity}
\usage{
measure_structuralcomplexity(A, norm = F)
}
\arguments{
\item{DSM}{A dependency structure matrix, reflecting the interfaces between elements.}

\item{nrom}{Boolean, default=F. If set to true, the normalized measure is returned.
\insertCite{Sinha.2016;textual}{EAD} proof that the maximum matrix energy is always \eqn{E_max<=n^{3/2}}}
}
\value{
The calculated structural complexity measure.
}
\description{
This measure is defined by \insertCite{Sinha.2018;textual}{EAD} and \insertCite{Sinha.2014;textual}{EAD}.
The C1 term is neglected here. Therefore only C2 and C3 are calculated. \code{SC=C2*C3*1/NROW(A)}.
Non-symmetric matrices are transformed into a symmetric binary matrix.
}
\examples{

DSM<-matrix(c(1,0,0,
            1,1,0,
            1,1,1),
         nrow=3,
         ncol=3,
         byrow=T)

measure_structuralcomplexity(DSM)
}
\references{
\insertAllCited{}
}
