% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_TOP10}
\alias{measure_TOP10}
\title{Measure the skewness of an vector}
\usage{
measure_TOP10(x)
}
\arguments{
\item{x}{A numeric vector which is strictly positive \code{0<x}.}
}
\value{
Returns the corresponding proportion
}
\description{
This function measures the skewness of an given input vector such as demand or resource costs.
It calculates the proportion of top ten percentage of elements on the total sum.
\insertCite{Anand2019;textual}{EAD} use this measure to operationalize resource cost distribution.
}
\examples{

data('csd_EAD')
measure_TOP10(CSD_EAD$DMD)
}
\references{
\insertAllCited{}
}
