% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{P}
\alias{P}
\title{Calculate the Product Line Commonality Index}
\usage{
P(P)
}
\arguments{
\item{P}{A product matrix representing the components per product (P_DD) with products in rows and components in columns.}
}
\value{
Returns the corresponding index value
}
\description{
This index was introduced by \insertCite{Kota2000;textual}{EAD}. The f-values as defined by Kota are set to one.
}
\examples{

P<-matrix(c(1,0,0,1,
              0,1,0,0,
              1,0,0,0,
              0,1,1,0,
              0,1,0,1),
       nrow = 4,
       ncol = 5)

measure_PCI(P)
}
\references{
\insertAllCited{}
}
