% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_Driver.R
\name{partmgmtCosts}
\alias{partmgmtCosts}
\title{Calculates Components' Part Administration Costs}
\usage{
partmgmtCosts(P_PD, DEMAND, C_PA)
}
\arguments{
\item{P_PD}{The product mix in the physical domain containing products in rows and components in columns.}

\item{DEMAND}{A numeric demand vector. If products are excluded from the product mix, the entries are zero.}

\item{PDC_fix}{A numeric vector holding the fixed costs for a component. These costs are calculated by tracing resource costs back into the physical domain.}

\item{R_PA}{Proportion of part administration costs on \code{PDC_fix}.}
}
\value{
Returns the summed development costs.
}
\description{
This function calculates components' part administration costs as defined by \insertCite{Meerschmidt.2024;textual}{EAD}.
}
\details{
For further details see \insertCite{Meerschmidt.2024;textual}{EAD}.
}
\examples{

require(EAD)
EAD <- smallEAD
P_PD <- EAD$P$PD
DEMAND<-c(10,50,100,30)
C_PA<-c(140,78,202,140)

development_costs(P_PD,
                  DEMAND,
                  PDC_fix,
                  C_PA)
}
\references{
\insertAllCited{}
}
