% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{pre_episodes}
\alias{pre_episodes}
\title{Identificação preliminar dos episódios de mudança de regime}
\usage{
pre_episodes(.data)
}
\arguments{
\item{.data}{dados do projeto \emph{"Varieties of Democracy"}.}
}
\description{
Identificação preliminar dos episódios de mudança de regime
}
