% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{pre_events}
\alias{pre_events}
\title{Discrimina eventos menores de estagnação e avanços, bem como eventos de reação.}
\usage{
pre_events(.data, .events, .backward_expr, .stagnant_expr, .forward_expr)
}
\arguments{
\item{.data}{dados do projeto \emph{"Varieties of Democracy"}.}

\item{.events}{define os eventos de avanço, estagnação ou reação.
\itemize{
\item \code{.forward}:
}

por definição \code{-0.0099 <= .forward <= -0.006} em episódios
de autocratização e \code{0.006 <= .forward <= 0.0099} em episódios de democratização.
\code{.forward} especifica variações menores, mas próximas ao \code{threshold}.
\itemize{
\item \code{.stagnant}:
}

por definição \code{-0.0059 <= .stagnant <= 0.0059} em episódios de
autocratização e de democratização. \code{.stagnant} especifica eventos de estagnação.
\itemize{
\item \code{.backward}:
}

por definição, \code{.backward < 0.0059} em episódios de autocratização
e \code{.backward < - 0.0059} em episódios de democratização. \code{.backward}, por fim,
indica eventos de reação.}

\item{.backward_expr}{expressão definida em \code{measures_exprs()} para a estimação
dos eventos de reação}

\item{.stagnant_expr}{expressão definida em \code{measures_exprs()} para a
estimação dos eventos de estagnação}

\item{.forward_expr}{expressão definida em \code{measures_exprs()} para a estimação
dos eventos de avanço}
}
\description{
Discrimina eventos menores de estagnação e avanços, bem como eventos de reação.
}
