% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RDU2}
\alias{RDU2}
\title{A DMM containing the architecture design of a power steering system in automobiles (RDU2) used by \insertCite{Jung.2022;textual}{EAD}.}
\format{
A DMM with 31 by 20 elements.
}
\usage{
data(RDU2)
}
\description{
A DMM containing the architecture design of a power steering system in automobiles (RDU2) used by \insertCite{Jung.2022;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
