% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{robot_DMM_control}
\alias{robot_DMM_control}
\title{A DMM containing inter-domain dependencies of a tree-axis robot control part used by \insertCite{Sawai.2017;textual}{EAD}.}
\format{
A DMM with 30 by 17 elements.
}
\usage{
data(robot_DMM_control)
}
\description{
A DMM containing inter-domain dependencies of a tree-axis robot control part used by \insertCite{Sawai.2017;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
