% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{robot_DMM_traverse}
\alias{robot_DMM_traverse}
\title{A DMM containing inter-domain dependencies of a tree-axis robot traverse part used by \insertCite{Sawai.2017;textual}{EAD}.}
\format{
A DMM with 32 by 26 elements.
}
\usage{
data(robot_DMM_traverse)
}
\description{
A DMM containing inter-domain dependencies of a tree-axis robot traverse part used by \insertCite{Sawai.2017;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
