% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{robot_DSM_control}
\alias{robot_DSM_control}
\title{A DSM containing component dependencies of a tree-axis robot control part used by \insertCite{Sawai.2017;textual}{EAD}.}
\format{
A DSM with 17 by 17 elements.
}
\usage{
data(robot_DSM_control)
}
\description{
A DSM containing component dependencies of a tree-axis robot control part used by \insertCite{Sawai.2017;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
