% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{robot_DSM_full}
\alias{robot_DSM_full}
\title{A DSM containing component dependencies of a tree-axis robot used by \insertCite{Sawai.2017;textual}{EAD}.}
\format{
A DSM with 151 by 151 elements.
}
\usage{
data(robot_DSM_full)
}
\description{
A DSM containing component dependencies of a tree-axis robot used by \insertCite{Sawai.2017;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
