% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{robot_DSM_traverse}
\alias{robot_DSM_traverse}
\title{A DSM containing component dependencies of a tree-axis robot traverse part used by \insertCite{Sawai.2017;textual}{EAD}.}
\format{
A DSM with 26 by 26 elements.
}
\usage{
data(robot_DSM_traverse)
}
\description{
A DSM containing component dependencies of a tree-axis robot traverse part used by \insertCite{Sawai.2017;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
