% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row.R
\name{row_update}
\alias{row_update}
\title{Faz pequenos ajustes na tipologia Regimes of the World}
\usage{
row_update(.data, ..., .index)
}
\arguments{
\item{.data}{Banco de dados}

\item{...}{Variáveis que devem ser especificadas
\itemize{
\item \code{.year} = série histórica
\item \code{.countries} = países
\item \code{.typology_d} = classificação dicotômica
\item \code{.typology_p} = classificação policotômica
\item \code{.reg_info} = nome do regime, quando ele começa e quando ele termina
\item \code{.reg_end_type} = nome do processo mais importante que resultou no
colapso do regime
}}

\item{.index}{Índice de democracia usado para estimar os episódios
de mudança de regime.
\itemize{
\item \code{IDE}: Índice de Democracia Eleitoral
\item \code{IDL}: Índice de Democracia Liberal
\item \code{IDI}: Índice de Democracia Igualitária
\item \code{IDP}: Índice de Democracia Participativa
\item \code{IDD}: Índice de Democracia Deliberativa
}

Por definição, eu adoto o Índice de Democracia Eleitoral (IDE), que mensura
os pré-resiquitos institucionais da democracia, conforme foi proposto por
Dahl (1971) e revisto em Dahl(1989), Dahl (1998) e em Coppedge et al.(2020).

IDE corresponde ao princípio, valor ou aspecto eleitoral da democracia.
Os demais índices agregam o Índice de Democracia Eleitoral a componentes
específicos da democracia. Por exemplo, o Índice de Democracia Liberal
equilave ao: IDE + componente liberal da democracia.

Para mais informações, consulte o guia do usuário do projeto v-DEM
(Coppedge et al., 2020) e o codebook do banco de dados (Coppedge et al., 2022).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Esta função flexibiliza o ponto de corte usado na codificação de
regimes políticos e desconsidera as transições que resultam de variações
anuais em e-DEM contidas no intervalo [-0.05; 0.05] para todos os
casos em que tais episódios sucedem imediatamente um retorno ao regime
anterior. Em outros termos, transições de regime com duração de um ano, fruto
de variações de 5\% em e-DEM são descartadas. Argumento que
essas transições devem ser tratadas como ruídos, pois o próprio conceito de
regime político pressupõe um mínimo de continuidade no tempo. (Schedler, 2013).
}
