% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_Driver.R
\name{setupCosts}
\alias{setupCosts}
\title{Calculates the setup costs}
\usage{
setupCosts(EAD, C_hold, C_order, R_setup, PrCU_var)
}
\arguments{
\item{EAD}{An EAD object created by \link[EAD]{crt_EAD}.}

\item{R_setup}{the proportion of setup costs on process manufacturing costs for a given lot.}

\item{PrCU_var}{The variable process unit costs.}

\item{R_hold}{the ratio of component's holding costs on total component costs.}

\item{R_order}{the ratio of component's order costs on total costs.}
}
\value{
Returns the summed setup costs as well as the lot sizes for the individual components
}
\description{
This function calculates the setup costs for a specific EAD design. The exact procedure is described in \insertCite{Meerschmidt.2024;textual}{EAD}.
See details for more information.
}
\details{
For further details see \insertCite{Meerschmidt.2024;textual}{EAD} as well as the sub models of \insertCite{Thonemann.2000;textual}{EAD} and \insertCite{Zhang.2020;textual}{EAD}.
A vignette is available by running \code{utils::vignette('setupCosts',package='EAD')}.
}
\examples{

require(EAD)
EAD <- smallEAD
EAD$DEMAND <- c(10,50,100,30)
R_hold <- rep(0.05,4)
R_order <- rep(0.15,4)
R_setup <- rep(0.2,4)
PrCU_var <- c(100,26.9,2.19,35.7)

clc_setupCosts(EAD,
               R_hold,
               R_order,
               R_setup,
               PrCU_var)
}
\references{
\insertAllCited{}
}
