% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{start_epi}
\alias{start_epi}
\title{Define o início do episódio de mudança de regime e
expande a série histórica em casos de descontinuidade.}
\usage{
start_epi(.data, .countries, .year, .index, .threshold, .start_epi_expr)
}
\arguments{
\item{.data}{dados do projeto \emph{"Varieties of Democracy"}.}

\item{.countries}{países}

\item{.year}{série histórica}

\item{.index}{Índice de democracia usado para estimar os episódios
de mudança de regime.
\itemize{
\item \code{IDE}: Índice de Democracia Eleitoral
\item \code{IDL}: Índice de Democracia Liberal
\item \code{IDI}: Índice de Democracia Igualitária
\item \code{IDP}: Índice de Democracia Participativa
\item \code{IDD}: Índice de Democracia Deliberativa
}

Por definição, eu adoto o Índice de Democracia Eleitoral (IDE), que mensura
os pré-resiquitos institucionais da democracia, conforme foi proposto por
Dahl (1971) e revisto em Dahl(1989), Dahl (1998) e em Coppedge et al.(2020).

IDE corresponde ao princípio, valor ou aspecto eleitoral da democracia.
Os demais índices agregam o Índice de Democracia Eleitoral a componentes
específicos da democracia. Por exemplo, o Índice de Democracia Liberal
equilave ao: IDE + componente liberal da democracia.

Para mais informações, consulte o guia do usuário do projeto v-DEM
(Coppedge et al., 2020) e o codebook do banco de dados (Coppedge et al., 2022).}

\item{.threshold}{Define o valor mínimo de variação no
índice de democracia para determinar o ínicio de processos de mudança de regime.
\code{.threshold} aceita valores entre 0 e 1.

Por definição, 0.01(1\%) em episódios de democratização e -0.01(-1\%) em
episódios de autocratização.}

\item{.start_epi_expr}{expressão definida em \code{measures_exprs()} usada
para estimar do início dos episódios de mudança de regime}
}
\description{
Define o início do episódio de mudança de regime e
expande a série histórica em casos de descontinuidade.
}
