% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trainbogie}
\alias{trainbogie}
\title{A DSM containing component dependencies of the train boogie used by \insertCite{Sinha.2018;textual}{EAD} and \insertCite{Kim.2016;textual}{EAD}.}
\format{
A DSM with 149 by 149 elements.
}
\usage{
data(trainbogie)
}
\description{
A DSM containing component dependencies of the train boogie used by \insertCite{Sinha.2018;textual}{EAD} and \insertCite{Kim.2016;textual}{EAD}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
