% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdem}
\alias{vdem}
\title{vDEM}
\format{
Lista de variáveis
\describe{
\item{country_name}{nome do país}
\item{year}{ano}
\item{regime}{Classificação ordinal de regimes políticos (RoW)}
\item{regime2}{Classificação dicotômica de regimes políticos (RoW)}
\item{reg_end_type}{Nome do processo mais importante que resultou no colapso do regime}
\item{reg_info}{Nome do regime, quando ele começa e quando ele termina}
\item{IDE}{Índide de Democracia Eleitoral}
\item{IDL}{Índide de Democracia Liberal}
\item{IDI}{Índide de Democracia Igualitária}
\item{IDP}{Índide de Democracia Participativa}
\item{IDD}{Índice de Democracia Deliberativa}
}
}
\source{
Varieties of Democracy (2022)
}
\usage{
vdem
}
\description{
O banco de dados \code{vdem} contém variáveis originais e tratadas do projeto
\emph{Varieties of Democracy}
}
\keyword{datasets}
