% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc_PCB.R
\name{clc_PCB}
\alias{clc_PCB}
\title{Calculate (true) Product Costs}
\usage{
clc_PCB(
  RES_CONS_PAT,
  DMD = NULL,
  RC_direct = NULL,
  RC_var = NULL,
  RC_fix = NULL,
  RCU = NULL,
  RCU_direct = NULL
)
}
\arguments{
\item{RES_CONS_PAT}{The resource consumption pattern matrix.}

\item{DMD}{A demand vector. If no demand vector is supplied, each product is assumed to have a demand of one.}

\item{RC_var}{A resource cost vector holding the variable costs.}

\item{RC_fix}{A resource cost vector holding the fixed costs.}

\item{RCU}{A vector holding the resource costs per unit for the variable indirect costs.}

\item{RCU_direct}{A vector holding the resource costs per unit for the variable direct costs.}
}
\value{
Calculates the true product costs (PC_B) and returns the resource costs per unit (RCU)
}
\description{
This function calculates the benchmark (true) product costs (PC_B).
There are two options to use this function:
\describe{
 \item{\code{!is.null(RC_var) & is.null(RCU)}}{This returns the product costs for a given EAD. \code{sum(PC_B*DMD)==TC}
 See the the first example for details.}
 \item{\code{is.null(RC_var) & !is.null(RCU)}}{This returns the product costs for an EAD with an reduced product mix. \code{sum(PC_B*DMD)<TC}
 See the second example for details.}
}
}
\examples{
### Calculate product costs for the designed EAD ###
data("exampleEAD")
TC =  sum(exampleEAD[[1]][[1]]$RC$fix +
          exampleEAD[[1]][[1]]$RC$var +
          exampleEAD[[1]][[1]]$RC$direct)

costs<- clc_PCB(RES_CONS_PAT = exampleEAD[[1]][[1]]$P$RD,
                DMD = exampleEAD[[1]][[1]]$DEMAND,
                RC_direct = exampleEAD[[1]][[1]]$RC$direct,
                RC_var = exampleEAD[[1]][[1]]$RC$var,
                RC_fix = exampleEAD[[1]][[1]]$RC$fix)

costs$PC_B

### Calculate product costs for an EAD with an reduced product mix
## lets assume we drop the first ten products

DMD<-exampleEAD[[1]][[1]]$DEMAND
DMD[1:10]<-0
costs<-clc_PCB(RES_CONS_PAT = exampleEAD[[1]][[1]]$P$RD,
               DMD = DMD,
               RC_direct = exampleEAD[[1]][[1]]$RC$direct,
               RC_fix = exampleEAD[[1]][[1]]$RC$fix,
               RCU=costs$RCU)
}
