% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_DEMAND.R
\name{crt_DEMAND}
\alias{crt_DEMAND}
\title{Creates the demand of products}
\usage{
crt_DEMAND(n_PROD, TOTAL_DEMAND, cv)
}
\arguments{
\item{n_PROD}{The number of products.}

\item{TOTAL_DEMAND}{the total demand the products in the portfolio.}

\item{cv}{Desired coefficient of variation as a vector representing the lower or upper bounds or as a double.}
}
\value{
Returns the demand vector and statistics
}
\description{
Creates a demand vector with a given level of heterogeneity which is specified by cv.
}
\examples{
# example
TOTAL_DEMAND<-1000
crt_DEMAND(n_PROD=120,TOTAL_DEMAND,cv=c(0.1,1))
}
