% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_EAD.R
\name{crt_EAD_MC}
\alias{crt_EAD_MC}
\title{Create EADs using a design of experiments (DOE)}
\usage{
crt_EAD_MC(
  DOE,
  uB_DMM = uB_DMM,
  ub_DSM = ub_DSM,
  allowZero = F,
  NUMB_CORES = 4,
  cluster = F,
  logfile = "",
  extMC_lib = F,
  ehNodes = "remove"
)
}
\arguments{
\item{DOE}{A data.frame, where each row contain a parameter set for an EAd realization. The data frame contain the following columns.
\describe{
  \item{N_FR}{The number of elements in the functional domain (FD)}
  \item{N_DD}{The number of elements in the physcial domain (PD)}
  \item{N_PrD}{The number of elements in the process domain (PrD)}
  \item{N_RD}{The number of elements in the resource domain (RD)}
  \item{DENS_FD}{Densitiy of the DSM_FD matrix.}
  \item{TOTAL_DEMAND}{Total demand}
  \item{DMD_cv}{Coefficient of variation for the demand vector.}
  \item{SDC_in}{System design complexity for the DMMs. In the current version this values is applied for all domains.}
  \item{ut_DMM}{Boolean. If \code{ut_DMM=TRUE}, the upper triangle DMMs are created representing indirect inter domain dependencies.}
  \item{SC_adj_n_in}{Adjusted and normalized structural complexity measure which should be created. In the current version this values is applied for all domains.}
  \item{TC}{NTotal costs}
  \item{ratio_fixedC}{Proportion of fixed costs on total costs}
  \item{RC_cor}{Correlation between variable cost vector and fixed cost vector}
  \item{RC_cv}{Coefficient of variation for resource cost distribution}
  \item{N_RUN}{Number of runs per unique setting}
}}

\item{NUMB_CORES}{Number of cores to run the simulation. Default: \code{NUMB_CORES=4}}

\item{logfile}{Optional argument, specifying the name of the logfile which is created in the working directory.}

\item{time_limit}{Time in seconds after which each EAD creation is interrupted. The interruption produces an error.
Therefore runs longer than \code{time_limit} are excluded from the output. The default value is infinity which means that no interruption is done.}
}
\value{
A list of EAD objects
}
\description{
This function creates EAD realizations using a design of experiments (DOE). The design is specified by \code{DOE}
}
\examples{
set.seed(1234)

N_FR=7
N_DD=15
N_PrD=20
N_RD=50
DENS_FD=c(0,0.05,0.1,0.2)
TOTAL_DEMAND=1000
DMD_cv=c(0,0.5,1,3)
DMM_PAR=c(0,0.05,0.1,0.2,0.3,0.4,0.5)
ut_DMM=c(F)
SC_adj_n_in=c(0,0.05,0.01,0.015,0.02)
N_RUN=1:1
x<-36
DOE<-expand.grid(list(N_FR = N_FR,
                      N_DD = N_DD,
                      N_PrD = N_PrD,
                      N_RD = N_RD,
                      DENS_FD = DENS_FD,
                      TOTAL_DEMAND = TOTAL_DEMAND,
                      DMD_cv = DMD_cv,
                      DMM_PAR = DMM_PAR,
                      DMM_method="SDC",
                      ut_DMM = ut_DMM,
                      SC_adj_n_in = SC_adj_n_in,
                      N_RUN = 1:N_RUN))

crt_EAD_MC(DOE,NUMB_CORES=4,logfile="log.txt")
}
