% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_RC.R
\name{crt_RC}
\alias{crt_RC}
\title{Create a resource cost object}
\usage{
crt_RC(N_RD, TC = 10^6, ratio_fixedC, RC_cor, RC_cv, max_tries = 15)
}
\arguments{
\item{N_RD}{Number of resources.}

\item{TC}{Total costs.}

\item{ratio_fixedC}{Proportion of fixed costs on total costs}

\item{RC_cor}{Correlation between variable cost vector and fixed cost vector}

\item{RC_cv}{Coefficient of variation for resource cost distribution}
}
\value{
A list containing the fixed and variable cost vector and the measured coefficient of variation for each cost vector as well as
the top 10% largest resource costs for each vector.
}
\description{
This function creates two resource cost vectors.
A fixed resource cost vector representing the total fixed costs as proportion of total costs defined by \code{TC*ratio_fixedC}.
Variable costs are given as \code{TC*(1-ratio_fixedC)}.
}
\examples{
set.seed(1234)

RC<-crt_RC(N_RD=50,TC=10^6,ratio_fixedC=0.4,RC_cor=0.5,RC_cv=0.2)

}
