% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_Driver.R
\name{development_costs}
\alias{development_costs}
\title{Calculates Components' Development  Costs}
\usage{
development_costs(
  P_PD,
  DEMAND,
  C_dvl,
  overdesign_Change = NULL,
  bounds = c(1, 1)
)
}
\arguments{
\item{P_PD}{The product mix in the physical domain containing products in rows and components in columns.}

\item{DEMAND}{A numeric demand vector. If products are excluded from the product mix, the entries are zero.}

\item{C_dvl}{The development costs for each component.}

\item{overdesign_Change}{A vector of length two, specifying the overdesign. The first entry indicates the index of the component which is substituted by the second indexed component.}

\item{bounds}{A vector specifying the bounds for a uniform distribution.}
}
\value{
Returns the summed development costs.
}
\description{
This function calculates components' development costs as defined by \insertCite{Meerschmidt.2024;textual}{EAD}.
}
\details{
For further details see \insertCite{Meerschmidt.2024;textual}{EAD}.
}
\examples{

require(EAD)
EAD <- smallEAD
P_PD <- EAD$P$PD
DEMAND<-c(10,50,100,30)
C_dvl<-c(210,117,303,210)

development_costs(P_PD,
                  DEMAND,
                  PDC_fix,
                  R_dvl)
}
\references{
\insertAllCited{}
}
