% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_designComplexity}
\alias{measure_designComplexity}
\title{Calculate Design Complexity}
\usage{
measure_designComplexity(A, norm = T)
}
\arguments{
\item{A}{A domain mapping matrix (DMM) or dependency structure matrix (DSM), reflecting the transition between two domains.}

\item{norm}{Boolean, default\code{TRUE}. Decides if the normalized design complexity measures should be returned.}
}
\value{
The calculated design complexity measure.
}
\description{
This measure is based on the Shannon entropy. Adapted from Modrak & Bednar (2015).
}
\examples{

A<-matrix(c(1,0,0,
            1,1,0,
            1,1,1),
         nrow=3,
         ncol=3,
         byrow=T)

measure_designComplexity(A)
}
