% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_HIC}
\alias{measure_HIC}
\title{Calculates the Interface Complexity (HIC)}
\usage{
measure_HIC(DSM)
}
\arguments{
\item{DSM}{A dependency structure matrix, reflecting the interfaces between elements.}
}
\value{
Returns the HIC value
}
\description{
This function calculates the Interface Complexity (HIC) as defined by \insertCite{Holtta.2005;textual}{EAD}.
Note, this function is not implemented yet since description of authors is not sufficient to reproduce the algorithm.
}
\examples{
require(EAD)

## water Sprinkler example by Ameri et al. (2008) ##
data("waterSprinkler")
DSM<-waterSprinkler

measure_HIC(DSM)
}
\references{
\insertAllCited{}
}
