% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_MCC}
\alias{measure_MCC}
\title{Calculates the McCabe's Cyclomatic Complexity (MCC)}
\usage{
measure_MCC(DSM)
}
\arguments{
\item{DSM}{A dependency structure matrix, reflecting the interfaces between elements.}
}
\value{
Returns the MCC value
}
\description{
This function calculates the McCabe's cyclomatic complexity (MCC)  \insertCite{McCabe.1976;textual}{EAD}.
}
\examples{
require(EAD)

## water Sprinkler example by Ameri et al. (2008) ##
data("waterSprinkler")
DSM<-waterSprinkler

measure_MCC(DSM)
}
\references{
\insertAllCited{}
}
