% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_modularity}
\alias{measure_modularity}
\title{measure_modularity}
\usage{
measure_modularity(A, preMember = NULL, plot_op = F)
}
\arguments{
\item{A}{A data.frame containing the dependency structure matrix with information on element dependencies.}

\item{preMember}{A numerical vector containing the assignment of vertices to cluster.
The default value is \code{preMember=NULL} which means that the communities are identified using the \link[igraph]{fastgreedy.community}}

\item{plot_op}{Boolean, default \code{plot_op=FALSE}. If set to true the graph is visualized.}
}
\value{
The calculated modularity measure
}
\description{
Calculates the directed of undirected modularity measure of a network.
The directed approach is descriped by Blondel, Guillaume, Lambiotte, & Lefebvre (2008) as well as Newman (2006).
The undirected measure is based on Leicht & Newmann (2008).
}
\examples{

## unsymmetrical matrix as input
## is transformed into a symmetric matrix first
## use the undirected approach by Blondel, Guillaume, Lambiotte, & Lefebvre (2008) as well as Newman (2006)
  DSM<-matrix(c(1,0,0,
              1,1,0,
              1,1,1),
           nrow=3,
           ncol=3,
           byrow=T)
  measure_modularity(DSM,symmetric=T)

## unsymmetrical matrix as input
## not transformed into a symmetric matrix first
## use the directed approach by Leicht & Newmann (2008)
  measure_modularity(DSM,symmetric=F)

}
