% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_reangularity}
\alias{measure_reangularity}
\title{Calculate Reangularity}
\usage{
measure_reangularity(A)
}
\arguments{
\item{DMM}{A domain mapping matrix, reflecting the transition between two domains.}
}
\value{
The calculated reangularity measure.
}
\description{
This measure calculates the reangularity for a given domain mapping matrix (DMM) according to Suh (1995) and Delaš et al. (2018).
}
\examples{

DMM<-matrix(c(1,0,0,
            1,1,0,
            1,1,1),
         nrow=3,
         ncol=3,
         byrow=T)

measure_reangularity(DMM)
}
