% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_semiangularity}
\alias{measure_semiangularity}
\title{Calculate Semangularity}
\usage{
measure_semiangularity(A)
}
\arguments{
\item{DMM}{A domain mapping matrix, reflecting the transition between two domains.}
}
\value{
The calculated semangularity measure.
}
\description{
This measure calculates the semangularity for a given domain mapping matrix (DMM) according to \insertCite{Suh.1995;textual}{EAD} and \insertCite{Delas.2018;textual}{EAD}.
}
\examples{

DMM<-matrix(c(1,0,0,
            1,1,0,
            1,1,1),
         nrow=3,
         ncol=3,
         byrow=T)

measure_semiangularity(DMM)
}
\references{
\insertAllCited{}
}
