% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityMeasures.R
\name{measure_structuralcomplexity}
\alias{measure_structuralcomplexity}
\title{Calculate structural complexity}
\usage{
measure_structuralcomplexity(A, norm = F)
}
\arguments{
\item{DSM}{A dependency structure matrix, reflecting the interfaces between elements.}
}
\value{
The calculated structural complexity measure.
}
\description{
This measure is defined by \insertCite{Sinha.2018;textual}{EAD} and \insertCite{Sinha.2014;textual}{EAD}.
Non-symmetric matrices are transformed into a symmetric binary matrix.
}
\examples{
library(EAD)

## proof of matrix energy
F2_left<-matrix(0,nrow=7,ncol=7)
F2_left[2:7,1]<-1
F2_left[1,2:7]<-1

sum(svd(F2_left)$d) # should be 4.90

F2_right<-matrix(0,nrow=7,ncol=7)
F2_right[1,2:3]<-1
F2_right[2,4:5]<-1
F2_right[3,6:7]<-1
F2_right<-makeMatrixsymmetric(F2_right)
sum(svd(F2_right)$d) # should be 6.83


DSM<-trainbogie

measure_structuralcomplexity(DSM)
}
\references{
\insertAllCited{}
}
