% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_Driver.R
\name{overdesign_EAD}
\alias{overdesign_EAD}
\title{Substitutes two components by an overdesigned one}
\usage{
overdesign_EAD(EAD)
}
\arguments{
\item{EAD}{An EAD object created via \link[EAD]{crt_EAD_MC}.}
}
\value{
Returns the manipulated EAD.
}
\description{
This function randomly chooses two component and replace those by one overdesigned one.
The first component is chosen randomly. The second component is selected by finding the nearest neighbor of component one.
Since components are substituted this function changes the following matrices \eqn{DMM_{FD,PD}}, \eqn{DSM_{PD}} and \eqn{DMM_{PD,PrD}}.
}
\details{
For further details see the corresponding vignette by running \code{utils::vignette(package = ‘EAD’)} and \insertCite{Meerschmidt.2024;textual}{EAD}
}
\examples{

EAD <- smallEAD
overdesign_EAD(EAD)
}
\references{
\insertAllCited{}
}
