% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row.R
\name{row_update}
\alias{row_update}
\title{Faz pequenos ajustes na tipologia Regimes of the World}
\usage{
row_update(.data, ...)
}
\arguments{
\item{.data}{Banco de dados}

\item{...}{Variáveis que devem ser especificadas
\itemize{
\item \code{.countries} = países
\item \code{typology_d} = ROW classifição dicotômica
\item \code{typology_p} = ROW classificação policotômica
\item \code{.index} = Índice de Democracia Eleitoral
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Esta função flexibiliza o ponto de corte usado na codificação de
regimes políticos e desconsidera as transições que resultam de variações
anuais no Índice de Democracia Eleitoral contidas no intervalo [-0.05; 0.05] para todos os
casos em que tais episódios sucedem imediatamente um retorno ao regime
anterior. Em outros termos, transições de regime com duração de um ano, fruto
de variações de 5\% no Índice de Democracia Eleitoral são descartadas. Argumento que
essas transições devem ser tratadas como ruídos, pois o próprio conceito de
regime político pressupõe um mínimo de continuidade no tempo. (Schedler, 2013).
}
